# Fix potential buffer overrun with heavily customized input and padding