Because number parsing will keep going as long as it can find new numbers, there is a case where it could trail off past SIMDJSON_PADDING:

```c++
const char *json = "{\"key\":122222222222222222222222222222222222222222222}";
parser.parse(json, strlen("{\"key\":1"));
```

We fix this for numbers we find at the root by copying the buffer and adding spaces; we fix it for arrays by checking if the array is closed. But we don't check if hashes are closed, so we can get into this situation still.

To make this a corruption bug, you just have to cut the array off after SIMDJSON_PADDING digits.

This closes off all possible corruption bugs where non-root primitives assume there is a structural or whitespace after them, I believe.

## Performance

Incredibly and inexplicably, this PR improves performance and reduces instruction counts on skylake:

| File                             | Blocks | master Cycles | PR Cycles | +Throughput | master Instr. | PR Instr. | -Instr. |
| ---                              |    --: |   --: |   --: |   --: |   --: |   --: |   --: |
| numbers                          |   2345 | 266.9 | 242.1 |    9% | 819.3 | 810.8 |    1% |
| mesh                             |  11306 | 308.9 | 302.2 |    2% | 1009.0 | 1010.5 |    0% |
| github_events                    |   1017 |  79.5 |  78.3 |    1% | 264.5 | 262.9 |    0% |
| canada                           |  35172 | 288.8 | 285.7 |    1% | 973.9 | 980.2 |    0% |
| random                           |   7976 | 141.6 | 140.2 |    1% | 494.6 | 491.8 |    0% |
| twitter                          |   9867 |  91.7 |  90.8 |    1% | 290.2 | 288.2 |    0% |
| update-center                    |   8330 | 112.0 | 111.0 |    0% | 335.5 | 332.4 |    0% |
| gsoc-2018                        |  51997 |  71.8 |  71.6 |    0% | 167.0 | 166.4 |    0% |
| apache_builds                    |   1988 |  85.9 |  85.8 |    0% | 304.7 | 302.9 |    0% |
| twitterescaped                   |   8787 | 184.5 | 184.4 |    0% | 489.9 | 487.7 |    0% |
| citm_catalog                     |  26987 |  80.2 |  80.3 |    0% | 296.3 | 296.3 |    0% |
| mesh.pretty                      |  24646 | 179.5 | 179.7 |    0% | 592.0 | 592.6 |    0% |
| instruments                      |   3442 | 105.2 | 105.9 |    0% | 378.5 | 378.3 |    0% |
| marine_ik                        |  46616 | 294.5 | 297.6 |   -1% | 919.6 | 917.5 |    0% |