@lemire 2020-08-12
(For people reading this, John has to go out of his way to trigger this issue by doing essentially undocumented tricks.)


@jkeiser There seems to be issue with VS16.... do you understand what the error is?