@lemire 2020-08-12
```
  D:\a\simdjson\simdjson\tests\errortests.cpp(153,19): error : static_assert expression is not an integral constant expression [D:\a\_temp\windows\tests\errortests.vcxproj]
205
  D:\a\simdjson\simdjson\tests\errortests.cpp(153,19): message : non-constexpr function 'strlen' cannot be used in a constant expression [D:\a\_temp\windows\tests\errortests.vcxproj]
206
  D:\a\simdjson\simdjson\tests\errortests.cpp(165,19): error : static_assert expression is not an integral constant expression [D:\a\_temp\windows\tests\errortests.vcxproj]
207
  D:\a\simdjson\simdjson\tests\errortests.cpp(165,19): message : non-constexpr function 'strlen' cannot be used in a constant expression [D:\a\_temp\windows\tests\errortests.vcxproj]
208
  D:\a\simdjson\simdjson\tests\errortests.cpp(176,19): error : static_assert expression is not an integral constant expression [D:\a\_temp\windows\tests\errortests.vcxproj]
209
  D:\a\simdjson\simdjson\tests\errortests.cpp(176,19): message : non-constexpr function 'strlen' cannot be used in a constant expression [D:\a\_temp\windows\tests\errortests.vcxproj]
```