@lemire 2020-08-12
@jkeiser Regarding performance, this should be, in the worse case, a one-time hit. And a tiny hit at that. Of course, compilers will disagree.

I am somewhat concerned about the need to access the end of the stream to solve this problem. I think it will come back to haunt us, but this PR does not add anything to the problem.