@jkeiser 2020-08-12
@lemire the puzzling thing is that this *increases* performance, not decreases :)