@lemire 2020-08-13
@jkeiser I could not verify these gains you report. Rather, I saw a small perf. decrease on small files.