@lemire 2020-08-13
AMD Rome, GCC 9. The numbers are not conclusive either way, but I would not call this as performance improvement. 

It is just fine (no regression to worry about).



Configuration
---------------
- Reference code: ref: refs/heads/master
- Candidate code: ref: refs/heads/jkeiser/number-corruption
Overall
---------------

|filename | ref. speed (GB/s) |candidate speed | difference |
|:--------| ----------:| --------------:| ----------:|
|google_maps_api_compact_response.json|1.090|1.095|0.5%|
|random.json|1.535|1.539|0.3%|
|marine_ik.json|0.819|0.820|0.2%|
|canada.json|0.782|0.782|0.0%|
|update-center.json|1.957|1.956|-0.1%|
|mesh.pretty.json|1.315|1.312|-0.2%|
|repeat.json|2.337|2.332|-0.2%|
|mesh.json|0.801|0.799|-0.3%|
|google_maps_api_response.json|2.116|2.110|-0.3%|
|apache_builds.json|2.528|2.516|-0.5%|
|twitter_api_response.json|2.537|2.525|-0.5%|
|twitterescaped.json|1.232|1.223|-0.7%|
|github_events.json|2.790|2.769|-0.8%|
|gsoc-2018.json|3.393|3.364|-0.8%|
|citm_catalog.json|2.643|2.621|-0.8%|
|instruments.json|2.119|2.100|-0.9%|
|numbers.json|0.902|0.893|-1.0%|
|twitter_api_compact_response.json|2.056|2.036|-1.0%|
|twitter.json|2.458|2.429|-1.2%|
|tree-pretty.json|2.225|2.197|-1.3%|
|twitter_timeline.json|1.681|1.647|-2.0%|

Variation: -2.0% -- 0.5%

|filename | ref. instructions/byte |candidate instructions/byte | difference |
|:--------| ----------:| --------------:| ----------:|
|canada.json|15.218|15.316|0.6%|
|instruments.json|5.916|5.938|0.4%|
|citm_catalog.json|4.630|4.638|0.2%|
|repeat.json|5.488|5.497|0.2%|
|mesh.json|15.713|15.736|0.1%|
|mesh.pretty.json|9.224|9.234|0.1%|
|tree-pretty.json|5.461|5.463|0.0%|
|google_maps_api_compact_response.json|10.763|10.763|0.0%|
|google_maps_api_response.json|5.675|5.675|0.0%|
|random.json|7.740|7.728|-0.2%|
|marine_ik.json|14.336|14.313|-0.2%|
|twitterescaped.json|7.673|7.660|-0.2%|
|gsoc-2018.json|2.614|2.609|-0.2%|
|twitter_api_compact_response.json|5.730|5.717|-0.2%|
|github_events.json|4.147|4.137|-0.2%|
|twitter.json|4.547|4.535|-0.3%|
|twitter_api_response.json|4.652|4.640|-0.3%|
|twitter_timeline.json|6.112|6.095|-0.3%|
|apache_builds.json|4.775|4.762|-0.3%|
|update-center.json|5.266|5.242|-0.5%|
|numbers.json|12.804|12.670|-1.0%|

Variation: -1.0% -- 0.6%