@jkeiser 2020-08-13
VS16 issue fixed.