@jkeiser 2020-08-18
I'll add last_structural. As for the dependence on the last character, I'm keenly aware and glad we've both got it in the back of our heads :) This just extends what we're already doing to hashes once I noticed this possible bug.