@lemire 2020-08-18
Merge at will.

Don't forget that we have https://simdjson.org/plots.html